web_scraping_intro
This is my first scraping code ever, made the basic web scraping of weather in different city from Google search. Tried using Beautiful Soup to do the scraping but it was kinda awful as there's a problem that i couldn't solve and i guess it was bc of the parsing thing so i continued with using the Requests_HTML instead.